requirements.txt
flask-sqlalchemy==2.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flask==1.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
pymysql==0.10.1
python-dotenv==0.14.0
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sortedcontainers==2.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sqlalchemy==1.3.20; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"